feat(venta): agregar venta atomica con idempotencia
La venta resuelve el precio vigente de cada producto, descuenta el stock y
calcula el total en una sola transaccion. Si falla la ultima linea por falta de
stock no queda nada a medias: ni mercaderia descontada ni una venta huerfana.
Cada linea congela el precio unitario, que es la razon por la que existe el
modulo de vigencias.

Dos decisiones que no se ven a simple vista:

- Las lineas se ordenan por id de producto antes de tocar el stock. Dos ventas
  con los mismos productos en distinto orden tomarian los FOR UPDATE en
  distinto orden y podrian trabarse mutuamente. Con un orden total y estable,
  la segunda espera a la primera en vez de abrazarse con ella.
- La idempotencia se apoya en la PK de clave_idempotencia, no en el chequeo
  previo. El chequeo cubre el caso normal (el POS reintenta segundos despues),
  pero dos reintentos simultaneos lo pasan los dos.

Se corrigen dos defectos encontrados por los tests:

- La escritura de la clave usaba save(), y sobre una entidad con id asignado
  Spring Data llama a merge() en vez de persist(): hace SELECT y luego UPDATE
  si la fila existe, con lo cual nunca viola la PK y el segundo escritor pisa
  al primero en silencio. Con 8 reintentos simultaneos se registraban las 8
  ventas y se descontaba stock 8 veces. Ahora es un INSERT nativo explicito.
- Instant maneja nanosegundos y timestamptz microsegundos, asi que la vigencia
  releida no coincidia con la asignada. Precio trunca al entrar, para que el
  objeto en memoria sea igual al persistido.

Se suman handlers para header obligatorio faltante y violaciones de constraint
en parametros, que caian en el handler generico y devolvian 500 en vez de 400.

+++ contratos/src/main/java/dev/francogomez/comercio/contratos/VentaConfirmada.java
@@ -30,6 +30,10 @@ public record VentaConfirmada(
 
     public static final int CURRENT_SCHEMA_VERSION = 1;
 
-    public record Linea(String sku, int cantidad, BigDecimal precioUnitario) {
+    /**
+     * La cantidad es decimal, no entera: el catálogo admite unidades de medida como el
+     * kilo, donde vender 0,750 es lo normal.
+     */
+    public record Linea(String sku, BigDecimal cantidad, BigDecimal precioUnitario) {
     }
 }
